Change to a unique endpoint for the sleep/sleep index predition + change to typing

+++ backend/Predictors.cs
@@ -2,11 +2,9 @@ namespace mlstack;
 
 public static class Predictors
 {
-    public static int SleepBDIPrediction(
-        double sleepQualityIndex,
-        int averageSleepHours)
+    //TODO
+    public static double SleepBDIPrediction(double sleepQualityIndex, int averageSleepHours)
     {
-        // Dummy prediction for now
         return 12345;
     }
 }
